+++ src/lib/mympd_state.h
@@ -71,6 +71,7 @@ struct t_mpd_features {
     bool pcre;                     //!< mpd supports pcre for filter expressions
     bool db_added;                 //!< mpd supports added attribute for songs
     bool sticker_sort_window;      //!< mpd supports sticker sort and window api
+    bool sticker_int;              //!< mpd supports sticker value handling as integer
     bool search_add_sort_window;   //!< mpd supports search and window for findadd/searchadd/searchaddpl
 };