+++ src/mpd_worker/smartpls.c
@@ -140,7 +140,7 @@ bool mpd_worker_smartpls_update(struct t_mpd_worker_state *mpd_worker_state, con
     }
 
     // get sort options
-    if (json_get_string(content, "$.sort", 0, 100, &sort, vcb_ismpdsort, NULL) == true &&
+    if (json_get_string(content, "$.sort", 0, 100, &sort, vcb_ismpd_sticker_sort, NULL) == true &&
         strcmp(sort, "shuffle") != 0)
     {
         json_get_bool(content, "$.sortdesc", &sortdesc, NULL);
@@ -161,7 +161,7 @@ bool mpd_worker_smartpls_update(struct t_mpd_worker_state *mpd_worker_state, con
     {
         if (json_get_string(content, "$.sticker", 1, NAME_LEN_MAX, &sds_buf1, vcb_isalnum, NULL) == true &&
             json_get_string(content, "$.value", 0, NAME_LEN_MAX, &sds_buf2, vcb_isname, NULL) == true &&
-            json_get_string(content, "$.op", 1, 2, &sds_buf3, vcb_iscompareop, NULL) == true)
+            json_get_string(content, "$.op", 1, 2, &sds_buf3, vcb_isstickerop, NULL) == true)
         {
             rc = mpd_worker_smartpls_update_sticker(mpd_worker_state, playlist, sds_buf1, sds_buf2, sds_buf3, sort, sortdesc, max_entries);
             if (rc == false) {
@@ -204,9 +204,9 @@ bool mpd_worker_smartpls_update(struct t_mpd_worker_state *mpd_worker_state, con
             rc = mpd_client_playlist_shuffle(mpd_worker_state->partition_state, playlist, NULL);
         }
         else if (strcmp(smartpltype, "sticker") == 0 &&
-                 sdslen(sort) > 0)
+                 sticker_sort_parse(sort) == MPD_STICKER_SORT_UNKOWN)
         {
-            // sticker based smart playlists are sorted by value as default, resort if needed
+            // resort sticker based smart playlists by sort tag
             rc = mpd_client_playlist_sort(mpd_worker_state->partition_state, playlist, sort, sortdesc, NULL);
         }
     }
@@ -385,15 +385,21 @@ static bool mpd_worker_smartpls_update_sticker(struct t_mpd_worker_state *mpd_wo
         return false;
     }
 
-    if (sort[0] != '\0' ||
+    enum mpd_sticker_sort sort_op = sticker_sort_parse(sort);
+
+    if (sort_op == MPD_STICKER_SORT_UNKOWN ||
         max_entries == 0)
     {
-        // we must get all entries and sort the result later
+        // we must get all entries
         max_entries = UINT_MAX;
     }
 
+    if (sort_op == MPD_STICKER_SORT_UNKOWN) {
+        sort_op = MPD_STICKER_SORT_URI;
+    }
+
     struct t_list *add_list = stickerdb_find_stickers_sorted(mpd_worker_state->stickerdb,
-        sticker, oper, value, MPD_STICKER_SORT_VALUE_INT, sort_desc, 0, max_entries);
+        sticker, oper, value, sort_op, sort_desc, 0, max_entries);
     if (add_list == NULL) {
         MYMPD_LOG_ERROR(NULL, "Could not fetch stickers for \"%s\"", sticker);
         return false;